==> 1_create_table_eur_usd_monthly.ipynb <==
import yfinance as yf
import os
import pandas as pd
from datetime import datetime

current_year = str(datetime.now().year)

### Télécharger les données historiques du EUR/USD (EURUSD=X)
ticker_eur_usd = "EURUSD=X"
df_eur_usd = yf.download(ticker_eur_usd, start=("2007-01-01"), end=current_year+"-12-31", auto_adjust=False)

df_eur_usd.columns = df_eur_usd.columns.droplevel(1)
df_eur_usd.reset_index(inplace=True)
df_eur_usd = df_eur_usd.drop(columns=["Adj Close", "Volume", "High", "Low", "Open"])

df_eur_usd["EUR_USD_monthly_avg"] = df_eur_usd["Close"].round(4)
df_eur_usd["Date"] = pd.to_datetime(df_eur_usd["Date"])
df_eur_usd = df_eur_usd.sort_values("Date")

# Garder la première valeur de chaque mois
df_eur_usd_monthly = (
    df_eur_usd
        .groupby(df_eur_usd["Date"].dt.to_period("M"))
        .first()
        .reset_index(drop=True)
)
# Mettre toutes les dates au premier du mois
df_eur_usd_monthly["Date_monthly"] = df_eur_usd_monthly["Date"].dt.to_period("M").dt.to_timestamp()
df_eur_usd_monthly = df_eur_usd_monthly.drop(columns=["Date", "Close"])

print(df_eur_usd_monthly)

os.makedirs("csv", exist_ok=True)
df_eur_usd_monthly.to_csv("csv/eur_usd_monthly.csv", index=False)
