==> 2_create_table_brent_eur_liter_monthly.ipynb <==
import yfinance as yf
import os
import pandas as pd
from datetime import datetime

current_year = str(datetime.now().year)

### Télécharger les données historiques du BRENT/USD (BZ=F)
ticker_brent = "BZ=F"
df_brent_usd = yf.download(ticker_brent, start=("2007-01-01"), end=current_year+"-12-31", auto_adjust=False)

df_brent_usd.columns = df_brent_usd.columns.droplevel(1)
df_brent_usd.reset_index(inplace=True)

df_brent_usd["Brent_USD_barrel"] = df_brent_usd["Close"].round(3)
df_brent_usd = df_brent_usd.drop(columns=["Adj Close", "Volume", "High", "Low", "Open", "Close"])

# Garder la première valeur de chaque mois
df_brent_usd_monthly = (
    df_brent_usd
        .groupby(df_brent_usd["Date"].dt.to_period("M"))
        .first()
        .reset_index(drop=True)
)

# Mettre toutes les dates au premier du mois
df_brent_usd_monthly["Date_monthly"] = df_brent_usd_monthly["Date"].dt.to_period("M").dt.to_timestamp()
df_brent_usd_monthly = df_brent_usd_monthly.drop(columns=["Date"])

print(df_brent_usd_monthly.head())




#### GET EUR USD by CSV
df_eur_usd_monthly_avg = pd.read_csv(
    "csv/eur_usd_monthly.csv",
    parse_dates=["Date_monthly"]
)

print("df_eur_usd_monthly_avg\n", df_eur_usd_monthly_avg)


# Merge monthly EUR/USD average with brent_usd_monthly
df_merged = pd.merge(df_brent_usd_monthly, df_eur_usd_monthly_avg, on='Date_monthly', how='left')
df_merged['brent_EUR_barrel'] = (df_merged['Brent_USD_barrel']/ df_merged['EUR_USD_monthly_avg']).round(5)

# 1 Brent Barrel contains 158,987 liters
df_merged["brent_EUR_liter"] = (df_merged["brent_EUR_barrel"] / 158.987).round(3)


# Apply a 1.1× coefficient to convert Brent price to Gazole price
df_merged["brent_gazole_EUR_liter"] = (df_merged["brent_EUR_liter"] * 1.1)

df_merged = df_merged.drop(columns=["brent_EUR_liter", "Brent_USD_barrel", "brent_EUR_barrel", "EUR_USD_monthly_avg"])


print('df_merged\n', df_merged)

os.makedirs("csv", exist_ok=True)
df_merged.to_csv("csv/brent_eur_liter_monthly.csv", index=False)
