==> 6_merge_all_datas_together.ipynb <==
import os
import pandas as pd

#### GET BRENT EUR by CSV
df_brent_eur = pd.read_csv(
    "csv/brent_eur_liter_monthly.csv",
    parse_dates=["Date_monthly"]
)
df_brent_eur.head(2)

#### GET REFINING MERGING BRENT EUR by CSV
df_refining_eur = pd.read_csv(
    "csv/refining_margin_brent_eur_liter_monthly.csv",
    parse_dates=["Date_monthly"]
)
df_refining_eur.head(2)

#### GET TRANSPORTATION FEES GAZOLE EUR by CSV
df_transport_fees_eur = pd.read_csv(
    "csv/transport_fees_eur_liter.csv",
    parse_dates=["Date_monthly"]
)
df_transport_fees_eur.head(2)

#### GET HTT TTC GAZOLE EUR by CSV
df_gazole_eur = pd.read_csv(
    "csv/gazole_htt_ttc_eur_liter_monthly.csv",
    parse_dates=["Date_monthly"]
)
df_gazole_eur.head(2)

# CREATE GLOBAL TABLE
df_merged = pd.merge(df_gazole_eur, df_brent_eur, on='Date_monthly', how='left')
df_merged = pd.merge(df_merged, df_refining_eur, on='Date_monthly', how='left')
df_merged = pd.merge(df_merged, df_transport_fees_eur, on='Date_monthly', how='left')
df_merged.head(5)

# VERIFY OFFICIAL HTT IS ACCURATE TO CALCULATE HTT
# HT = Brent + refining_margin + transportation_fees
df_merged["calculate_gazole_eur_liter_htt"] = (
df_merged["brent_gazole_EUR_liter"] + df_merged["refining_margin_brent_eur_liter"] + df_merged["Gazole_transportation_fees_eur_liter"]
)

# Create special col for gazole taxes
df_merged["gazole_taxes_eur_liter"] = df_merged["gazole_eur_liter_ttc"] - df_merged["gazole_eur_liter_htt"]
df_merged.head(5)

# Export CSV
os.makedirs("csv", exist_ok=True)
df_merged.to_csv("csv/df_table_all_datas.csv", index=False)
